add common component, homepage, login and register page
