+++ index.html
@@ -8,7 +8,7 @@
   <title>Fix Tech | Online Home Appliance Repair & Maintenance Service</title>
 
   <!-- Favicon -->
-  <link rel="shortcut icon" href="/img/favicon.png">
+  <link rel="shortcut icon" href="/img/favicon-2.png">
 
   <!-- Bootstrap CSS -->
   <link rel="stylesheet" href="/css/bootstrap.min.css">
