+++ public/css/style.css
@@ -317,7 +317,7 @@ h6 {
 	3. General
 -----------------------*/
 body {
-  font-family: "Fira Sans", sans-serif;
+  font-family: "Roboto", sans-serif;
   font-size: 16px;
   color: #676767;
   background: #FCFBFB;
@@ -3550,7 +3550,7 @@ table .badge {
 }
 @media (max-width: 991.96px) {
   .banner-slider {
-    background: #F2F7F6;
+    background-position: bottom;
     padding: 70px 0 70px;
   }
 }
@@ -10443,9 +10443,13 @@ table .badge {
   font-size: 14px;
 }
 .testimonial-item.testimonial-item-two:hover {
-  background: #111111;
+  background: #676767;
 }
 .testimonial-item.testimonial-item-two:hover .user-info h6 {
+  color: #fff;
+}
+
+.testimonial-item.testimonial-item-two:hover .user-info p {
   color: #fff;
 }
 
@@ -11444,7 +11448,8 @@ table .badge {
 }
 @media (max-width: 575.96px) {
   .login-wrapper .loginbox .span-or {
-    left: 0;
+    left: 50%;
+    transform: translateX(-50%);
   }
 }
 .login-wrapper .loginbox .span-or-log {
@@ -11459,7 +11464,8 @@ table .badge {
 }
 @media (max-width: 575.96px) {
   .login-wrapper .loginbox .span-or-log {
-    left: 0;
+    left: 50%;
+    transform: translateX(-50%);
   }
 }
 .login-wrapper .loginbox .lock-user {
