+++ src/components/common/Footer.jsx
@@ -1,0 +1,207 @@
+function Footer() {
+    return (
+        <>
+            <footer className="footer">
+                {/* Footer Top */}
+                <div className="footer-top aos" data-aos="fade-down">
+                    <div className="container">
+                        <div className="row">
+                            <div className="col-lg-7">
+                                <div className="row">
+                                    <div className="col-lg-4 col-md-6">
+                                        {/* Footer Widget */}
+                                        <div className="footer-widget footer-menu">
+                                            <h5 className="footer-title">About Company</h5>
+                                            <ul>
+                                                <li>
+                                                    <a href="about.html">Our Company</a>
+                                                </li>
+                                                <li>
+                                                    <a href="javascript:void(0)">Shop Toyota</a>
+                                                </li>
+                                                <li>
+                                                    <a href="javascript:void(0)">Dreamsrentals USA</a>
+                                                </li>
+                                                <li>
+                                                    <a href="javascript:void(0)">Dreamsrentals Worldwide</a>
+                                                </li>
+                                                <li>
+                                                    <a href="javascript:void(0)">Dreamsrental Category</a>
+                                                </li>
+                                            </ul>
+                                        </div>
+                                        {/* /Footer Widget */}
+                                    </div>
+                                    <div className="col-lg-4 col-md-6">
+                                        {/* Footer Widget */}
+                                        <div className="footer-widget footer-menu">
+                                            <h5 className="footer-title">Vehicles Type</h5>
+                                            <ul>
+                                                <li>
+                                                    <a href="javascript:void(0)">All Vehicles</a>
+                                                </li>
+                                                <li>
+                                                    <a href="javascript:void(0)">SUVs</a>
+                                                </li>
+                                                <li>
+                                                    <a href="javascript:void(0)">Trucks</a>
+                                                </li>
+                                                <li>
+                                                    <a href="javascript:void(0)">Cars</a>
+                                                </li>
+                                                <li>
+                                                    <a href="javascript:void(0)">Crossovers</a>
+                                                </li>
+                                            </ul>
+                                        </div>
+                                        {/* /Footer Widget */}
+                                    </div>
+                                    <div className="col-lg-4 col-md-6">
+                                        {/* Footer Widget */}
+                                        <div className="footer-widget footer-menu">
+                                            <h5 className="footer-title">Quick links</h5>
+                                            <ul>
+                                                <li>
+                                                    <a href="javascript:void(0)">My Account</a>
+                                                </li>
+                                                <li>
+                                                    <a href="javascript:void(0)">Champaigns</a>
+                                                </li>
+                                                <li>
+                                                    <a href="javascript:void(0)">Dreamsrental Dealers</a>
+                                                </li>
+                                                <li>
+                                                    <a href="javascript:void(0)">Deals and Incentive</a>
+                                                </li>
+                                                <li>
+                                                    <a href="javascript:void(0)">Financial Services</a>
+                                                </li>
+                                            </ul>
+                                        </div>
+                                        {/* /Footer Widget */}
+                                    </div>
+                                </div>
+                            </div>
+                            <div className="col-lg-5">
+                                <div className="footer-contact footer-widget">
+                                    <h5 className="footer-title">Contact Info</h5>
+                                    <div className="footer-contact-info">
+                                        <div className="footer-address">
+                                            <span>
+                                                <i className="feather-phone-call"></i>
+                                            </span>
+                                            <div className="addr-info">
+                                                <a href="tel:+1(888)7601940">+ 1 (888) 760 1940</a>
+                                            </div>
+                                        </div>
+                                        <div className="footer-address">
+                                            <span>
+                                                <i className="feather-mail"></i>
+                                            </span>
+                                            <div className="addr-info">
+                                                <a href="mailto:info@example.com">[email&#160;protected]</a>
+                                            </div>
+                                        </div>
+                                        <div className="update-form">
+                                            <form action="#">
+                                                <span>
+                                                    <i className="feather-mail"></i>
+                                                </span>
+                                                <input type="email" className="form-control" placeholder="Enter Your Email Here" />
+                                                <button type="submit" className="btn btn-subscribe">
+                                                    <span>
+                                                        <i className="feather-send"></i>
+                                                    </span>
+                                                </button>
+                                            </form>
+                                        </div>
+                                    </div>
+                                    <div className="footer-social-widget">
+                                        <ul className="nav-social">
+                                            <li>
+                                                <a href="javascript:void(0)">
+                                                    <i className="fa-brands fa-facebook-f fa-facebook fi-icon"></i>
+                                                </a>
+                                            </li>
+                                            <li>
+                                                <a href="javascript:void(0)">
+                                                    <i className="fab fa-instagram fi-icon"></i>
+                                                </a>
+                                            </li>
+                                            <li>
+                                                <a href="javascript:void(0)">
+                                                    <i className="fab fa-behance fi-icon"></i>
+                                                </a>
+                                            </li>
+                                            <li>
+                                                <a href="javascript:void(0)">
+                                                    <i className="fab fa-twitter fi-icon"></i>
+                                                </a>
+                                            </li>
+                                            <li>
+                                                <a href="javascript:void(0)">
+                                                    <i className="fab fa-linkedin fi-icon"></i>
+                                                </a>
+                                            </li>
+                                        </ul>
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+                {/* /Footer Top */}
+
+                {/* Footer Bottom */}
+                <div className="footer-bottom">
+                    <div className="container">
+                        {/* Copyright */}
+                        <div className="copyright">
+                            <div className="row align-items-center">
+                                <div className="col-md-6">
+                                    <div className="copyright-text">
+                                        <p>© 2024 Dreams Rent. All Rights Reserved.</p>
+                                    </div>
+                                </div>
+                                <div className="col-md-6">
+                                    {/* Copyright Menu */}
+                                    <div className="copyright-menu">
+                                        <div className="vistors-details">
+                                            <ul className="d-flex">
+                                                <li>
+                                                    <a href="javascript:void(0)">
+                                                        <img className="img-fluid" src="/img/icons/paypal.svg" alt="Paypal" />
+                                                    </a>
+                                                </li>
+                                                <li>
+                                                    <a href="javascript:void(0)">
+                                                        <img className="img-fluid" src="/img/icons/visa.svg" alt="Visa" />
+                                                    </a>
+                                                </li>
+                                                <li>
+                                                    <a href="javascript:void(0)">
+                                                        <img className="img-fluid" src="/img/icons/master.svg" alt="Master" />
+                                                    </a>
+                                                </li>
+                                                <li>
+                                                    <a href="javascript:void(0)">
+                                                        <img className="img-fluid" src="/img/icons/applegpay.svg" alt="applegpay" />
+                                                    </a>
+                                                </li>
+                                            </ul>
+                                        </div>
+                                    </div>
+                                    {/* /Copyright Menu */}
+                                </div>
+                            </div>
+                        </div>
+                        {/* /Copyright */}
+                    </div>
+                </div>
+                {/* /Footer Bottom */}
+            </footer>
+        </>
+    )
+}
+
+export default Footer;
