+++ src/index.css
@@ -1,0 +1,19 @@
+.font-14 {
+    font-size: 14px;
+}
+
+.banner-font {
+    font-size: 3.5rem;
+}
+
+.main-color {
+    color: #FFA633 !important; 
+}
+
+.mtd-30 {
+    margin-top: -30px;
+}
+
+.mb-20 {
+    margin-bottom: 20px !important;
+}
